+++ apps/hud/ui/overlays/track_radar/track_radar.py
@@ -25,123 +25,14 @@
 import logging
 import math
 from pathlib import Path
-from typing import Any, Dict, NamedTuple, Optional, final
-
-from PySide6.QtCore import QSize, Qt
-from PySide6.QtGui import QColor, QImage, QPainter, QRadialGradient
-from PySide6.QtQml import QQmlImageProviderBase
-from PySide6.QtQuick import QQuickImageProvider
+from typing import Any, Dict, Optional, final
 
 from apps.hud.ui.infra.hf_types import DriverMotionInfo, LiveSessionMotionInfo
 from apps.hud.ui.overlays.base import BaseOverlayQML
 from lib.config import OverlayId, OverlayPosition
 
-# -------------------------------------- CONSTANTS ---------------------------------------------------------------------
-
-class _CarDims(NamedTuple):
-    width_m: float
-    length_m: float
-
-# Real-world car dimensions keyed by formula type string
-_CAR_DIMENSIONS_M: dict[str, _CarDims] = {
-    "F1 Modern":      _CarDims(2.00, 5.63),
-    "F1 Classic":     _CarDims(2.00, 5.63),
-    "F1 Generic":     _CarDims(2.00, 5.63),
-    "F1 World":       _CarDims(2.00, 5.63),
-    "F1 Elimination": _CarDims(2.00, 5.63),
-    "Esports":        _CarDims(2.00, 5.63),
-    "F2":             _CarDims(1.90, 5.285),
-    "F2 2021":        _CarDims(1.90, 5.285),
-}
-_DEFAULT_CAR_DIMENSIONS_M = _CarDims(2.00, 5.63)
-
-# Radar display constants — single source of truth for both Python and QML.
-# QML reads radarAreaRatio via a property set in post_setup so the canvas
-# geometry and the Python coordinate projection always agree.
-_RADAR_RANGE_M    = 25.0   # metres represented by half the radar area
-_RADAR_BASE_WIDTH = 300    # must match baseWidth in track_radar.qml
-_RADAR_AREA_RATIO = 0.85   # must match the 0.85 factor in track_radar.qml
-_RADAR_AREA_PX    = _RADAR_BASE_WIDTH * _RADAR_AREA_RATIO
-
-def _car_px(formula_type: str) -> tuple[float, float]:
-    """Return (width_px, length_px) scaled to the radar coordinate system."""
-    dims = _CAR_DIMENSIONS_M.get(formula_type, _DEFAULT_CAR_DIMENSIONS_M)
-    px_per_m = (_RADAR_AREA_PX / 2.0) / _RADAR_RANGE_M
-    return round(dims.width_m * px_per_m, 1), round(dims.length_m * px_per_m, 1)
-
-
-def _make_glow_image(is_left: bool, car_w_px: float, car_l_px: float) -> QImage:
-    """
-    Generate a single _RADAR_BASE_WIDTH x _RADAR_BASE_WIDTH glow sector image for the left or right side.
-
-    Replicates the Canvas atan2-based pie sector with a radial gradient:
-      centre → 0.8 alpha, 30% out → 0.4 alpha, edge → 0.0 alpha.
-    The image is generated once at base resolution; scaleFactor is handled
-    by the GPU Scale transform on the parent Item.
-    """
-    size = _RADAR_BASE_WIDTH
-    half_r = _RADAR_AREA_PX / 2.0
-    cx = cy = size / 2.0
-    half_w = car_w_px / 2.0
-    half_l = car_l_px / 2.0
-
-    img = QImage(size, size, QImage.Format.Format_ARGB32_Premultiplied)
-    img.fill(QColor(0, 0, 0, 0))
-
-    painter = QPainter(img)
-    painter.setRenderHint(QPainter.RenderHint.Antialiasing)
-
-    grad = QRadialGradient(cx, cy, half_r)
-    grad.setColorAt(0.0,  QColor(255, 0, 0, int(0.8 * 255)))
-    grad.setColorAt(0.3,  QColor(255, 0, 0, int(0.4 * 255)))
-    grad.setColorAt(1.0,  QColor(255, 0, 0, 0))
-    painter.setBrush(grad)
-    painter.setPen(Qt.PenStyle.NoPen)
-
-    # Mirror the Canvas atan2 angles exactly.
-    # Canvas: left glow arc from angleTopRight to angleBottomRight (clockwise).
-    #   angleTopRight    = atan2(-halfL,  halfW)   →  upper-right of car = left side of radar
-    #   angleBottomRight = atan2( halfL,  halfW)   →  lower-right of car
-    # Canvas: right glow arc from angleBottomLeft to angleTopLeft (clockwise).
-    #   angleTopLeft     = atan2(-halfL, -halfW)
-    #   angleBottomLeft  = atan2( halfL, -halfW)
-    # QPainter.drawPie uses 1/16° units, 0° = 3 o'clock, counter-clockwise positive.
-    # Convert from Canvas (radians, CW from 3 o'clock) to QPainter (1/16°, CCW from 3 o'clock).
-    rect_x = int(cx - half_r)
-    rect_y = int(cy - half_r)
-    rect_size = int(half_r * 2)
-
-    # Both wedges are the same angular width; just differ in which side of the circle.
-    # half_angle is the half-angle of the narrow wedge (e.g. ~70° for F1 car dims).
-    half_angle_deg = math.degrees(math.atan2(half_l, half_w))
-
-    if is_left:
-        # Narrow wedge centred on 0° (3 o'clock = right side of radar)
-        qt_start = int(-half_angle_deg * 16)
-        qt_span  = int( half_angle_deg * 2 * 16)
-    else:
-        # Narrow wedge centred on 180° (9 o'clock = left side of radar)
-        qt_start = int((180 - half_angle_deg) * 16)
-        qt_span  = int( half_angle_deg * 2 * 16)
-
-    painter.drawPie(rect_x, rect_y, rect_size, rect_size, qt_start, qt_span)
-    painter.end()
-    return img
-
-
-class _RadarGlowImageProvider(QQuickImageProvider):
-    """Serves pre-baked left/right glow sector images to QML via image://radar/glow-left|glow-right."""
-
-    def __init__(self, car_w_px: float, car_l_px: float):
-        super().__init__(QQmlImageProviderBase.ImageType.Image)
-        self._images = {
-            "glow-left":  _make_glow_image(True,  car_w_px, car_l_px),
-            "glow-right": _make_glow_image(False, car_w_px, car_l_px),
-        }
-
-    def requestImage(self, image_id: str, size: QSize, requested_size: QSize) -> QImage:  # pylint: disable=unused-argument
-        return self._images.get(image_id, QImage())
-
+from ._glow_provider import RadarGlowImageProvider
+from ._radar_math import _RADAR_AREA_RATIO, _RADAR_RANGE_M, car_px, to_radar_coords
 
 # -------------------------------------- CLASSES -----------------------------------------------------------------------
 
@@ -176,8 +67,8 @@ class TrackRadarOverlay(BaseOverlayQML):
     @final
     def pre_setup(self):
         """Register the glow image provider before QML loads."""
-        car_w_px, car_l_px = _car_px("F1 Modern")
-        self.qml_engine.addImageProvider("radar", _RadarGlowImageProvider(car_w_px, car_l_px))
+        car_w_px, car_l_px = car_px("F1 Modern")
+        self.qml_engine.addImageProvider("radar", RadarGlowImageProvider(car_w_px, car_l_px))
 
     @final
     def post_setup(self):
@@ -220,7 +111,7 @@ class TrackRadarOverlay(BaseOverlayQML):
         if not ref_driver or not ref_driver.car_motion:
             return
 
-        car_w_px, car_l_px = _car_px(data.formula_type)
+        car_w_px, car_l_px = car_px(data.formula_type)
         self.set_qml_property("carWidthPx", car_w_px)
         self.set_qml_property("carLengthPx", car_l_px)
 
@@ -253,7 +144,7 @@ class TrackRadarOverlay(BaseOverlayQML):
             if self._is_car_on_right(rel_x, rel_z):
                 car_on_right = True
 
-            radar_x, radar_y = self._to_radar_coords(rel_x, rel_z)
+            radar_x, radar_y = to_radar_coords(rel_x, rel_z)
             # Preserve pre-rewrite visual heading convention from QML:
             # old delegate used `rotation: -(driver.heading || 0)`.
             car_data.extend([radar_x, radar_y, -d['heading'], dist <= _RADAR_RANGE_M])
@@ -267,15 +158,6 @@ class TrackRadarOverlay(BaseOverlayQML):
     @staticmethod
     def _is_car_on_right(rel_x: float, rel_z: float) -> bool:
         return 1.5 < rel_x < 4.0 and abs(rel_z) < 8.0
-
-    @staticmethod
-    def _to_radar_coords(rel_x: float, rel_z: float) -> tuple[float, float]:
-        # Match legacy QML projection exactly:
-        #   radarArea.center + local offset where radarArea was inset in 300x300 root.
-        # This resolves to root center while keeping radarArea-sized scaling.
-        center = _RADAR_BASE_WIDTH / 2
-        scale = (_RADAR_AREA_PX / 2) / _RADAR_RANGE_M
-        return center - rel_x * scale, center - rel_z * scale
 
     def _get_reference_driver(self, session: LiveSessionMotionInfo) -> Optional[DriverMotionInfo]:
         """Get the reference driver from session data."""
